# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Home Stock Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ItemStatus = typeof ItemStatus[keyof typeof ItemStatus];


export const ItemStatus = {
  in_stock: 'in_stock',
  low: 'low',
  out: 'out',
} as const;

export interface Product {
  id: number;
  barcode: string;
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Item {
  id: number;
  name: string;
  category: string;
  location: string;
  status: ItemStatus;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  minThreshold?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  updatedBy?: string | null;
  /** @nullable */
  expirationDate?: string | null;
  /** @nullable */
  productId?: number | null;
  product?: Product | null;
  createdAt: string;
  updatedAt: string;
}

export type ItemInputStatus = typeof ItemInputStatus[keyof typeof ItemInputStatus];


export const ItemInputStatus = {
  in_stock: 'in_stock',
  low: 'low',
  out: 'out',
} as const;

export interface ItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  location?: string;
  status: ItemInputStatus;
  quantity?: string;
  unit?: string;
  minThreshold?: number;
  notes?: string;
  updatedBy?: string;
  expirationDate?: string;
  /** @nullable */
  productId?: number | null;
}

export type ItemUpdateStatus = typeof ItemUpdateStatus[keyof typeof ItemUpdateStatus];


export const ItemUpdateStatus = {
  in_stock: 'in_stock',
  low: 'low',
  out: 'out',
} as const;

export interface ItemUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  location?: string;
  status?: ItemUpdateStatus;
  quantity?: string;
  unit?: string;
  minThreshold?: number;
  notes?: string;
  updatedBy?: string;
  expirationDate?: string;
  /** @nullable */
  productId?: number | null;
}

export type StatusUpdateStatus = typeof StatusUpdateStatus[keyof typeof StatusUpdateStatus];


export const StatusUpdateStatus = {
  in_stock: 'in_stock',
  low: 'low',
  out: 'out',
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
  updatedBy?: string;
}

export interface QuantityAdjust {
  /** Amount to add (positive) or subtract (negative) from current quantity */
  delta: number;
}

export interface ItemsSummary {
  total: number;
  inStock: number;
  low: number;
  out: number;
}

export interface CategoryCount {
  category: string;
  total: number;
  inStock: number;
  low: number;
  out: number;
}

export interface LocationCount {
  location: string;
  total: number;
  inStock: number;
  low: number;
  out: number;
}

export interface ShoppingItem {
  id: number;
  name: string;
  checked: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ShoppingItemInput {
  /** @minLength 1 */
  name: string;
}

export interface ShoppingItemUpdate {
  /** @minLength 1 */
  name?: string;
  checked?: boolean;
}

export interface ProductInput {
  /** @minLength 1 */
  barcode: string;
  /** @minLength 1 */
  name: string;
  brand?: string;
  category?: string;
  imageUrl?: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  brand?: string;
  category?: string;
  imageUrl?: string;
}

export type ListItemsParams = {
search?: string;
category?: string;
location?: string;
status?: ListItemsStatus;
};

export type ListItemsStatus = typeof ListItemsStatus[keyof typeof ListItemsStatus];


export const ListItemsStatus = {
  in_stock: 'in_stock',
  low: 'low',
  out: 'out',
} as const;

export type ListProductsParams = {
search?: string;
};